requirements.txt
maturin==1.7.1
numpy==1.24.4
opencv-python==4.10.0.84
pillow==10.4.0